Restructure UT: move setup away
