+++ common/tests/cfg_ut.rs
@@ -1,34 +1,15 @@
+mod ut_rt;
+
 #[cfg(test)]
 mod cfg_v1_ut {
-    use core::panic;
-    use flakes::config::{
-        self,
-        cfgparse::FlakeCfgParser,
-        itf::{self, FlakeConfig, InstanceMode},
-    };
-    use std::{env, path::PathBuf};
-
-    /// Setup the test
-    fn setup(cfg_path: String) -> Option<FlakeConfig> {
-        if let Ok(parser) = FlakeCfgParser::new(env::current_dir().unwrap().join("tests").join("data").join(&cfg_path), vec![]) {
-            return parser.parse();
-        }
-
-        None
-    }
-
-    /// Run a test bundle
-    fn tb<T>(cfg_path: String, probe: T) -> ()
-    where
-        T: Fn(Option<FlakeConfig>) -> () + panic::UnwindSafe,
-    {
-        probe(setup(cfg_path));
-    }
+    use super::ut_rt;
+    use flakes::config::itf::InstanceMode;
+    use std::path::PathBuf;
 
     /// Test Firecracker configuration v1 overall parse
     #[test]
     fn test_cfg_v1_fc_overall_parse() {
-        tb("cfg-v1/firecracker.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/firecracker.yaml".to_string(), |cfg| {
             assert!(cfg.is_some(), "FlakeConfig v1 for firecracker should not be None");
             assert!(cfg.unwrap().version() == 1, "Version should be 1");
         });
@@ -37,7 +18,7 @@ mod cfg_v1_ut {
     /// Test bogus configuration v42 overall parse
     #[test]
     fn test_cfg_v42_overall_parse() {
-        tb("bogus.yaml".to_string(), |cfg| {
+        ut_rt::tb("bogus.yaml".to_string(), |cfg| {
             assert!(cfg.is_none(), "FlakeConfig v42 should be None and must be unsupported");
         });
     }
@@ -45,7 +26,7 @@ mod cfg_v1_ut {
     /// Test podman configuration v1 overall parse
     #[test]
     fn test_cfg_v1_pdm_overall_parse() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             assert!(cfg.is_some(), "FlakeConfig v1 for podman should not be None");
             assert!(cfg.unwrap().version() == 1, "Version should be 1");
         });
@@ -54,7 +35,7 @@ mod cfg_v1_ut {
     /// Test OCI name
     #[test]
     fn test_cfg_v1_pdm_name() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.runtime().image_name() == "banana");
         });
@@ -63,7 +44,7 @@ mod cfg_v1_ut {
     /// Test OCI target path
     #[test]
     fn test_cfg_v1_pdm_exported_app_path() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.runtime().paths().exported_app_path() == &PathBuf::from("/banana/in/the/container"));
         });
@@ -72,7 +53,7 @@ mod cfg_v1_ut {
     /// Test OCI target path
     #[test]
     fn test_cfg_v1_pdm_registered_app_path() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.runtime().paths().registered_app_path() == &PathBuf::from("/usr/bin/banana"));
         });
@@ -81,7 +62,7 @@ mod cfg_v1_ut {
     /// Test OCI base layer
     #[test]
     fn test_cfg_v1_pdm_base_layer() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.runtime().base_layer().is_some(), "Base layer should be defined");
             assert!(cfg.runtime().base_layer().unwrap() == "cobol_rules");
@@ -91,7 +72,7 @@ mod cfg_v1_ut {
     /// Test OCI additional layers
     #[test]
     fn test_cfg_v1_pdm_layers() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.runtime().layers().is_some(), "There should be more than one additional layers");
             assert!(cfg.runtime().layers().unwrap().len() == 2);
@@ -105,7 +86,7 @@ mod cfg_v1_ut {
     /// Test OCI container needs to be run as user root (UID 0)
     #[test]
     fn test_cfg_v1_pdm_run_as() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.runtime().run_as().is_some(), "User should be defined");
             assert!(cfg.runtime().run_as().unwrap().uid.is_root(), "User should be root");
@@ -116,7 +97,7 @@ mod cfg_v1_ut {
     /// Test OCI should be resumed and attached
     #[test]
     fn test_cfg_v1_pdm_mode_flags() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!((*cfg.runtime().instance_mode() & InstanceMode::Attach) == InstanceMode::Attach, "Should have attach flag");
             assert!((*cfg.runtime().instance_mode() & InstanceMode::Resume) == InstanceMode::Resume, "Should have resume flag");
@@ -126,7 +107,7 @@ mod cfg_v1_ut {
     /// Test OCI target podman args
     #[test]
     fn test_cfg_v1_pdm_args() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.engine().args().is_some(), "Podman should run with some parameters");
         });
@@ -135,7 +116,7 @@ mod cfg_v1_ut {
     /// Test OCI target podman args examination
     #[test]
     fn test_cfg_v1_pdm_args_exm() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let args = cfg.unwrap().engine().args().unwrap();
             assert!(args.len() == 3, "Podman should have three params");
             assert!(args == ["--storage-opt size=10G", "--rm", "-ti"], "Podman should have parameters in a certain order");
@@ -145,7 +126,7 @@ mod cfg_v1_ut {
     /// Test OCI includes
     #[test]
     fn test_cfg_v1_pdm_static() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(cfg.static_data().get_bundles().is_some(), "Podman wants to include something");
         });
@@ -154,7 +135,7 @@ mod cfg_v1_ut {
     /// Test OCI includes contains a specific archive
     #[test]
     fn test_cfg_v1_pdm_static_data() {
-        tb("cfg-v1/podman.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v1/podman.yaml".to_string(), |cfg| {
             let cfg = cfg.unwrap();
             assert!(
                 cfg.static_data().get_bundles().unwrap() == ["irq-dropout.tar.gz"],
@@ -167,7 +148,7 @@ mod cfg_v1_ut {
     /// Test v2 overall parse
     #[test]
     fn test_cfg_v2_overall_parse() {
-        tb("cfg-v2/all.yaml".to_string(), |cfg| {
+        ut_rt::tb("cfg-v2/all.yaml".to_string(), |cfg| {
             assert!(cfg.is_some(), "FlakeConfig v2 should not be None");
             assert!(cfg.unwrap().version() == 2, "Version should be 2");
         });
